fix: convert URI components to instances in fake file-system also
