+++ packages/app/src/platform/electron/electron-preload/initialize-privileged-platform-modules.ts
@@ -1,18 +1,7 @@
-import type * as codeOSSFiles from '@pkerschbaum/code-oss-file-service/out/vs/platform/files/common/files';
 import { clipboard, ipcRenderer } from 'electron';
 
 import { VSBuffer } from '@app/base/buffer';
-import type { Event } from '@app/base/event';
-import type {
-  FileDeleteOptions,
-  IFileStat,
-  IFileStatWithMetadata,
-  IResolveFileOptions,
-  IResolveMetadataFileOptions,
-  IWatchOptions,
-} from '@app/base/files';
-import type { IDisposable } from '@app/base/lifecycle';
-import type { CoordinationArgs } from '@app/base/resources';
+import { fileServiceUriInstancesToComponents, PlatformFileService } from '@app/base/fileService';
 import { URI, UriComponents } from '@app/base/uri';
 import { config } from '@app/config';
 import { bootstrapDiskFileService } from '@app/platform/electron/electron-preload/bootstrap-disk-file-service';
@@ -65,37 +54,6 @@ type Privileged = {
   };
 };
 
-export type PlatformFileService = {
-  resolve(
-    resource: UriComponents,
-    options: IResolveMetadataFileOptions,
-  ): Promise<IFileStatWithMetadata>;
-  resolve(resource: UriComponents, options?: IResolveFileOptions): Promise<IFileStat>;
-  del(
-    resource: UriComponents,
-    options?: Partial<Omit<FileDeleteOptions, 'useTrash'>>,
-  ): Promise<void>;
-  copy(
-    source: UriComponents,
-    target: UriComponents,
-    overwrite?: boolean,
-    coordinationArgs?: CoordinationArgs,
-  ): Promise<IFileStatWithMetadata>;
-  move(
-    source: UriComponents,
-    target: UriComponents,
-    overwrite?: boolean,
-    coordinationArgs?: CoordinationArgs,
-  ): Promise<IFileStatWithMetadata>;
-  createFolder(resource: UriComponents): Promise<IFileStatWithMetadata>;
-  watch(resource: UriComponents, options?: IWatchOptions): IDisposable;
-  onDidFilesChange: Event<PlatformFileChangesEvent>;
-};
-
-type PlatformFileChangesEvent = {
-  affects(resource: UriComponents): boolean;
-};
-
 const CLIPBOARD_FILELIST_FORMAT = `${config.productName}/file-list`;
 
 export function initializePrivilegedPlatformModules() {
@@ -104,55 +62,7 @@ export function initializePrivilegedPlatformModules() {
   window.privileged = {
     // eslint-disable-next-line node/no-process-env
     processEnv: process.env,
-    fileService: {
-      async resolve(resource, options) {
-        const result = await fileService.resolve(uriComponentsToInstance(resource), options);
-        fileStatUriInstancesToComponents(result);
-        return result as unknown as any;
-      },
-      del(resource, options) {
-        return fileService.del(uriComponentsToInstance(resource), options);
-      },
-      async copy(source, target, overwrite, coordinationArgs) {
-        const result = await fileService.copy(
-          uriComponentsToInstance(source),
-          uriComponentsToInstance(target),
-          overwrite,
-          coordinationArgs,
-        );
-        fileStatUriInstancesToComponents(result);
-        return result;
-      },
-      async move(source, target, overwrite, coordinationArgs) {
-        const result = await fileService.move(
-          uriComponentsToInstance(source),
-          uriComponentsToInstance(target),
-          overwrite,
-          coordinationArgs,
-        );
-        fileStatUriInstancesToComponents(result);
-        return result;
-      },
-      async createFolder(resource) {
-        const result = await fileService.createFolder(uriComponentsToInstance(resource));
-        fileStatUriInstancesToComponents(result);
-        return result;
-      },
-      watch(resource, options) {
-        return fileService.watch(uriComponentsToInstance(resource), options);
-      },
-      onDidFilesChange(listener, ...delegated) {
-        return fileService.onDidFilesChange(
-          (e) =>
-            listener({
-              affects(resource) {
-                return e.affects(uriComponentsToInstance(resource));
-              },
-            }),
-          ...delegated,
-        );
-      },
-    },
+    fileService: fileServiceUriInstancesToComponents(fileService),
     app: {
       getPath: (...args) => ipcRenderer.invoke(APP_CHANNEL.GET_PATH, ...args),
     },
@@ -209,19 +119,4 @@ function bufferToResources(buffer: Uint8Array): UriComponents[] {
 
 function resourcesToBuffer(resources: UriComponents[]): Uint8Array {
   return VSBuffer.fromString(resources.map((r) => r.toString()).join('\n')).buffer;
-}
-
-function uriComponentsToInstance(uri: UriComponents) {
-  // eslint-disable-next-line no-restricted-syntax -- in this file we have the "boundary" to `@pkerschbaum/code-oss-file-service`, so it is allowed to use `URI.from` here
-  return URI.from(uri);
-}
-
-function fileStatUriInstancesToComponents<T extends codeOSSFiles.IFileStat>(stat: T): void {
-  (stat as { -readonly [prop in keyof IFileStat]: IFileStat[prop] }).resource =
-    stat.resource.toJSON();
-  if (stat.children) {
-    for (const child of stat.children) {
-      fileStatUriInstancesToComponents(child);
-    }
-  }
 }
