+++ packages/app/src/platform/file-system.types.ts
@@ -1,6 +1,6 @@
 import type { FileDeleteOptions } from '@app/base/files';
+import type { PlatformFileService } from '@app/base/fileService';
 import type { UriComponents } from '@app/base/uri';
-import type { PlatformFileService } from '@app/platform/electron/electron-preload/initialize-privileged-platform-modules';
 
 export type PlatformFileSystem = PlatformFileService & {
   trash(
